refactor(mlir): convert a function definition to its dispatch attribute
The kind a `sol.func` is reached through comes from the definition rather than
from a match at the emission site, since a regular function needs the node id
the kind alone does not carry.

+++ solx-mlir/src/context/function/mod.rs
@@ -86,7 +86,7 @@ impl<'context> Function<'context> {
             FunctionDispatch::Identifier(identifier) => {
                 operation_builder.id(IntegerAttribute::new(
                     IntegerType::new(context.melior, solx_utils::BIT_LENGTH_X64 as u32).into(),
-                    identifier as i64,
+                    usize::from(identifier) as i64,
                 ))
             }
             FunctionDispatch::Kind(function_kind) => {
